+++ Units/parser-kotlin.r/kotlin-destructuring.d/input.kt
@@ -1,0 +1,4 @@
+val (a, b) = functionReturningPair()
+val (c: Int, d: String) = functionReturningPair()
+var (e, f) = functionReturningPair()
+var (g: Int, h: Pair<String, Int>) = functionReturningPair()