+++ Units/parser-kotlin.r/kotlin-extensions.d/input.kt
@@ -1,0 +1,6 @@
+class A
+
+fun A.test() = true
+
+val A.value
+	get() = "AAA"